refactor: move some util functions
