+++ utils/project.ts
@@ -7,3 +7,17 @@ export const getGroupedProjects = (projects: Project[], project: Project) =>
   project.groupId
     ? projects.filter(({ groupId }) => groupId === project.groupId)
     : [project];
+
+export const isProjectEqual = (projectA: Project, projectB: Project) =>
+  projectA.title === projectB.title && projectA.subtitle === projectB.subtitle;
+
+export const includesProject = (
+  project: Project,
+  projectsToCheck: Project[],
+): boolean =>
+  !!projectsToCheck.find((checkedProject) =>
+    isProjectEqual(checkedProject, project),
+  );
+
+export const findProjectIndex = (project: Project, projects: Project[]) =>
+  projects.findIndex((p) => isProjectEqual(p, project));
