+++ utils/walkthroughs.ts
@@ -2,7 +2,6 @@ import { RoleId } from '@/types/roles';
 import { roles } from '../data/roles';
 import { walkthroughsByRole } from '../data/walkthroughs';
 import { GetWalkthroughScenario, Walkthrough } from '../types/walkthrough';
-import { Project } from '../types/project';
 
 const SCENARIO_ID_DELIMITER = '-';
 
@@ -112,17 +111,3 @@ export const getNextScenarioId = (scenarioId: string): string | undefined => {
 
   return createScenarioId(roleId, walkthroughIndex, nextScenarioIndex);
 };
-
-export const isProjectEqual = (projectA: Project, projectB: Project) =>
-  projectA.title === projectB.title && projectA.subtitle === projectB.subtitle;
-
-export const includesProject = (
-  project: Project,
-  projectsToCheck: Project[],
-): boolean =>
-  !!projectsToCheck.find((checkedProject) =>
-    isProjectEqual(checkedProject, project),
-  );
-
-export const findProjectIndex = (project: Project, projects: Project[]) =>
-  projects.findIndex((p) => isProjectEqual(p, project));
